Switch to storing each entry per file
It's more convenient and unix-way

+++ fzf.sh
@@ -1,35 +1,37 @@
 #!/bin/bash
 
-mkfifo /tmp/asdd
-mkfifo /tmp/res_to_copy
+RCLIP_HOME="$HOME/.rclip"
+
+cd $RCLIP_HOME
+
+PICKED_FILE=$(gawk '
+BEGIN {
+    RS = ""
+    FS ="\n"
+}
+BEGINFILE {
+    printf("%s:", FILENAME);
+};
+{
+    for (f=1; f<=NF; ++f) {printf("%s ", $f)};
+};
+ENDFILE {
+    printf("\n")
+};
+' $(rg --sort path --files-with-matches .) | fzf --tac --no-sort -d : --with-nth 2.. --preview "cat {1}" | awk -F : '{print $1}')
+#./UTF8_STRING/1619123453031
+
+if [[ -z "$PICKED_FILE" ]]; then
+    exit 1
+fi
+
+TARGET_NAME=$(dirname $PICKED_FILE)
+FILE_NAME="$RCLIP_HOME/$PICKED_FILE"
 
 # send a signal to rclip that now we will set entry from history
 pkill -SIGUSR1 ^rclip$
 
-# will wait until something will be printed to res_to_copy, and then pipe it to
-# xclip. We need to do this BEFORE run rclip, because named pipe res_to_copy
-# must be ready to read before we will write something to it.
-copy_func() {
-    read TARGET_NAME
-    FILE_INPUT=""
-    if [[ $TARGET_NAME == "! "* ]];
-    then
-        TARGET_NAME="${TARGET_NAME:2}"
-        FILE_INPUT=~/.rclip/other_targets/$TARGET_NAME/$(cat)
-    fi
+# nohup need to leave process running in the background (useful when call
+# script by hotkey)
+nohup xclip -t $TARGET_NAME -i $FILE_NAME -sel c > /dev/null 2> /dev/null
 
-    # nohup need to leave process running in the background (useful when call
-    # script by hotkey)
-    nohup xclip -i $FILE_INPUT -t $TARGET_NAME -sel c > /dev/null 2> /dev/null
-}
-copy_func < /tmp/res_to_copy &
-
-# set asdd as stdin for rclip (which will be used for get picked index of entry)
-# redirect rclip stderr to res_to_copy (which will be used for
-#    getting original text entry)
-# redirect rclip stdout to fzf
-# redirect fzf stdout to asdd (which used for send index of picked entry to rclip)
-rclip list_and_set < /tmp/asdd 2> /tmp/res_to_copy | fzf --tac --with-nth 2.. > /tmp/asdd
-
-rm /tmp/asdd
-rm /tmp/res_to_copy
